애니메이션, 기능
1. 몬스터 등장 히트 사망 아이들 애니메이션 적용
2. 게임종료
3. 게임정지
4. 상점 버튼 구현
5. 전투흐름 구현
6. 전투 횟수에 따른 적군 등장 빈도
7. 패배시 체력 회복 및 귀환

+++ SeeOfFools/Assets/Script/Stat.cs
@@ -6,8 +6,8 @@ public class Stat
 {
     public EnemyType type { get; }
     public string name { get; set; }
-    public int maxHp { get; set; }
-    public int curhp { get; set; }
+    public float maxHp { get; set; }
+    public float curhp { get; set; }
     public int Damage { get; set; }
     public int Gold { get; set; }
     public int Score { get; set; }
